+++ packages/solid/src/icons/Info.tsx
@@ -1,0 +1,23 @@
+import { Component, ComponentProps } from 'solid-js';
+
+const Info: Component<ComponentProps<'svg'>> = (props) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      {...props}
+    >
+      <circle cx="12" cy="12" r="9" stroke="black" stroke-width="2"/>
+<path d="M11 10C11 9.44772 11.4477 9 12 9V9C12.5523 9 13 9.44772 13 10V16C13 16.5523 12.5523 17 12 17V17C11.4477 17 11 16.5523 11 16V10Z" fill="#141313"/>
+<path d="M11 7C11 6.44772 11.4477 6 12 6V6C12.5523 6 13 6.44772 13 7V7C13 7.55228 12.5523 8 12 8V8C11.4477 8 11 7.55228 11 7V7Z" fill="#141313"/>
+    </svg>
+  );
+};
+export default Info;