+++ packages/vue/src/icons/Focus.ts
@@ -1,0 +1,22 @@
+import { h, type FunctionalComponent, type SVGAttributes } from 'vue';
+
+const Focus: FunctionalComponent<SVGAttributes> = (props, context) => {
+  const { attrs } = context;
+  return h('svg', {
+    xmlns: 'http://www.w3.org/2000/svg',
+    width: 24,
+    height: 24,
+    viewBox: '0 0 24 24',
+    fill: 'none',
+    stroke: 'currentColor',
+    'stroke-width': 2,
+    'stroke-linecap': 'round',
+    'stroke-linejoin': 'round',
+    ...attrs,
+    ...props,
+    innerHTML: "<path d='M2 7V5C2 3.34315 3.34315 2 5 2H7M2 17V19C2 20.6569 3.34315 22 5 22H7M17 22H19C20.6569 22 22 20.6569 22 19V17M22 7V5C22 3.34315 20.6569 2 19 2H17' stroke='black' stroke-width='2' stroke-linecap='round'/><circle cx='12' cy='12' r='3' stroke='black' stroke-width='2'/>"
+  });
+};
+
+Focus.displayName = 'Focus';
+export default Focus;