+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @melospot/icons-react
 
+## 0.0.4
+
+### Patch Changes
+
+- fix: ensure all icons are synced and published
+- Updated dependencies
+  - @melospot/icons@0.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- feat: add bin icon and sync all packages
+- Updated dependencies
+  - @melospot/icons@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes